Develop

See merge request nab/sel2b/admin-site!340

+++ src/app/pages/financial/ar-invoice/invoice.component.ts
@@ -35,6 +35,7 @@ export class InvoiceComponent implements OnInit {
     public onoffFilter: any;
     public listMoreFilter: any = [];
     searchForm: FormGroup;
+    public dateType = 0;
 
     constructor(public router: Router,
         public fb: FormBuilder,
@@ -62,11 +63,14 @@ export class InvoiceComponent implements OnInit {
         //  Assign get list function name, override letiable here
         this.tableService.getListFnName = 'getList';
         this.tableService.context = this;
+        this.keyService.watchContext.next(this);
     }
 
     ngOnInit() {
         //  Init Fn
-        this.listMoreFilter = [{ value: false, name: 'Date Filter' }];
+        // this.listMoreFilter = [{ value: false, name: 'Date Filter' }];
+        this.listMaster['listFilter'] = [{ value: false, name: 'Date Filter' }];
+        this.listMaster['dateType'] = [{ id: 0, name: 'Issue Date' }, { id: 1, name: 'Due Date' }];
         this.listMaster['status'] = [
             { id: 1, name: 'New' },
             { id: 2, name: 'Submitted' },
@@ -96,6 +100,26 @@ export class InvoiceComponent implements OnInit {
         setTimeout(() => {
             this.renderer.invokeElementMethod(this.inp.nativeElement, 'focus');
         }, 300);
+    }
+
+    onFilterChanged(value) {
+        if (!value) {
+            this.searchForm.patchValue({
+                'inv_dt_from': null,
+                'inv_dt_to': null,
+                'inv_due_dt_from': null,
+                'inv_due_dt_to': null
+            });
+        }
+    }
+
+    onDateTypeChanged() {
+        this.searchForm.patchValue({
+            'inv_dt_from': null,
+            'inv_dt_to': null,
+            'inv_due_dt_from': null,
+            'inv_due_dt_to': null
+        });
     }
 
     countInvoiceStatus() {
@@ -128,6 +152,33 @@ export class InvoiceComponent implements OnInit {
         });
     }
 
+    createInvoice() {
+        this.router.navigate(['/financial/invoice/create']);
+    }
+
+    viewInvoice(id?) {
+        if (id) {
+            this.router.navigate(['/financial/invoice/view', id]);
+        } else {
+            const selectedInvoiceId = this.list.items[this.selectedIndex].id;
+            if (selectedInvoiceId) {
+                this.router.navigate(['/financial/invoice/view', selectedInvoiceId]);
+            }
+        }
+    }
+
+    editInvoice(id?) {
+        if (id) {
+            this.router.navigate(['/financial/invoice/edit', id]);
+        } else {
+            const selectedInvoiceId = this.list.items[this.selectedIndex].id;
+            const selectedInvoiceStatus = this.list.items[this.selectedIndex].invoice_status_id;
+            if (selectedInvoiceId && selectedInvoiceStatus === 1) {
+                this.router.navigate(['/financial/invoice/edit', selectedInvoiceId]);
+            }
+        }
+    }
+
     getListInvoiceItemsRef() {
         this.financialService.getListInvoiceItemsRef().subscribe(res => {
             try {
@@ -156,20 +207,24 @@ export class InvoiceComponent implements OnInit {
         return stt.name;
     }
 
-    deleteInvoice(id) {
-        const modalRef = this.modalService.open(ConfirmModalContent);
-        modalRef.result.then(result => {
-            if (result) {
-                this.financialService.deleteInvoice(id).subscribe(res => {
-                    try {
-                        this.toastr.success(res.message);
-                        this.getList();
-                    } catch (e) {
-                        console.log(e);
-                    }
-                });
-            }
-        }, dismiss => { });
+    cancelInvoice(id?) {
+        const selectedInvoiceId = id ? id : this.list.items[this.selectedIndex].id;
+        const selectedInvoiceStatus = this.list.items[this.selectedIndex].invoice_status_id;
+        if (selectedInvoiceId && selectedInvoiceStatus === 1) {
+            const modalRef = this.modalService.open(ConfirmModalContent);
+            modalRef.result.then(result => {
+                if (result) {
+                    this.financialService.deleteInvoice(id).subscribe(res => {
+                        try {
+                            this.toastr.success(res.message);
+                            this.getList();
+                        } catch (e) {
+                            console.log(e);
+                        }
+                    });
+                }
+            }, dismiss => { });
+        }
     }
 
 }
